add calc.py
